+++ crates/cli/src/commands/alias.rs
@@ -1,4 +1,5 @@
 use crate::error::ProtoCliError;
+use crate::helpers::{map_pin_type, PinOption};
 use crate::session::ProtoSession;
 use clap::Args;
 use proto_core::{is_alias_name, Id, ProtoConfig, UnresolvedVersionSpec};
@@ -16,11 +17,11 @@ pub struct AliasArgs {
     #[arg(required = true, help = "Version or alias to associate with")]
     spec: UnresolvedVersionSpec,
 
-    #[arg(
-        long,
-        help = "Add to the global ~/.proto/.prototools instead of local ./.prototools"
-    )]
+    #[arg(long, group = "pin", help = "Add to the global ~/.proto/.prototools")]
     global: bool,
+
+    #[arg(long, group = "pin", help = "Location of .prototools to add to")]
+    to: Option<PinOption>,
 }
 
 #[tracing::instrument(skip_all)]
@@ -40,16 +41,20 @@ pub async fn alias(session: ProtoSession, args: AliasArgs) -> AppResult {
 
     let tool = session.load_tool(&args.id).await?;
 
-    let config_path = ProtoConfig::update(tool.proto.get_config_dir(args.global), |config| {
-        let tool_configs = config.tools.get_or_insert(Default::default());
+    let config_path = ProtoConfig::update(
+        tool.proto
+            .get_config_dir(map_pin_type(args.global, args.to)),
+        |config| {
+            let tool_configs = config.tools.get_or_insert(Default::default());
 
-        tool_configs
-            .entry(tool.id.clone())
-            .or_default()
-            .aliases
-            .get_or_insert(Default::default())
-            .insert(args.alias.clone(), args.spec.clone());
-    })?;
+            tool_configs
+                .entry(tool.id.clone())
+                .or_default()
+                .aliases
+                .get_or_insert(Default::default())
+                .insert(args.alias.clone(), args.spec.clone());
+        },
+    )?;
 
     println!(
         "Added alias {} ({}) to config {}",