+++ crates/cli/tests/unalias_test.rs
@@ -39,13 +39,11 @@ mod unalias_local {
         })
         .unwrap();
 
-        sandbox.debug_files();
-
         sandbox
             .run_bin(|cmd| {
                 cmd.arg("unalias").arg("node").arg("example");
             })
-            .debug();
+            .success();
 
         let config = load_config(sandbox.path());
 
@@ -119,6 +117,43 @@ mod unalias_global {
             .success();
 
         let config = load_config(sandbox.path().join(".proto"));
+
+        assert!(config.tools.get("node").unwrap().aliases.is_empty());
+    }
+}
+
+mod unalias_user {
+    use super::*;
+
+    #[test]
+    fn removes_existing_alias() {
+        let sandbox = create_empty_proto_sandbox();
+
+        ProtoConfig::update(sandbox.path().join(".home"), |config| {
+            config.tools.get_or_insert(Default::default()).insert(
+                Id::raw("node"),
+                PartialProtoToolConfig {
+                    aliases: Some(BTreeMap::from_iter([(
+                        "example".into(),
+                        UnresolvedVersionSpec::parse("19.0.0").unwrap(),
+                    )])),
+                    ..Default::default()
+                },
+            );
+        })
+        .unwrap();
+
+        sandbox
+            .run_bin(|cmd| {
+                cmd.arg("unalias")
+                    .arg("node")
+                    .arg("example")
+                    .arg("--from")
+                    .arg("user");
+            })
+            .success();
+
+        let config = load_config(sandbox.path().join(".home"));
 
         assert!(config.tools.get("node").unwrap().aliases.is_empty());
     }