+++ crates/cli/tests/utils.rs
@@ -29,12 +29,19 @@ impl Deref for ProtoSandbox {
 }
 
 fn apply_settings(sandbox: &mut Sandbox) {
+    let root = sandbox.path().to_path_buf();
+    let home_dir = sandbox.path().join(".home");
     let proto_dir = sandbox.path().join(".proto");
+
+    // Folders must exist or tests fail!
+    fs::create_dir_all(&home_dir).unwrap();
+    fs::create_dir_all(&proto_dir).unwrap();
 
     let mut env = HashMap::new();
     env.insert("RUST_BACKTRACE", "1");
     env.insert("WASMTIME_BACKTRACE_DETAILS", "1");
     env.insert("NO_COLOR", "1");
+    env.insert("PROTO_SANDBOX", root.to_str().unwrap());
     env.insert("PROTO_HOME", proto_dir.to_str().unwrap());
     env.insert("PROTO_LOG", "trace");
     env.insert("PROTO_TEST", "true");