+++ CHANGELOG.md
@@ -14,7 +14,16 @@
 
 #### 🚀 Updates
 
+- Added support for updating the `~/.prototools` file (root of user home directory).
+- Added `--pin=user` to `proto install`.
+- Added `--to=global|local|user` to `proto alias`, `pin`, and `plugin add`.
+- Added `--from=global|local|user` to `proto unalias`, `unpin`, and `plugin remove`.
+- Added aliases for pin locations.
+  - `cwd` -> `local`
+  - `home` -> `user`
+  - `store` -> `global`
 - Added new `settings.offline` settings that control how offline checks work.
+- Deprecated the `--global` flag.
 
 #### ⚙️ Internal